BlockRedstoneTorch added, extends BlockTorch
