+++ src/main/java/net/glowstone/block/ItemTable.java
@@ -114,8 +114,8 @@ public final class ItemTable {
         reg(Material.BED_BLOCK, new BlockBed());
         reg(Material.TORCH, new BlockTorch(Material.TORCH));
         reg(Material.REDSTONE_WIRE, new BlockRedstoneDust());
-        reg(Material.REDSTONE_TORCH_OFF, new BlockTorch(Material.REDSTONE_TORCH_OFF));
-        reg(Material.REDSTONE_TORCH_ON , new BlockTorch(Material.REDSTONE_TORCH_OFF));
+        reg(Material.REDSTONE_TORCH_OFF, new BlockRedstoneTorch(Material.REDSTONE_TORCH_OFF));
+        reg(Material.REDSTONE_TORCH_ON , new BlockRedstoneTorch(Material.REDSTONE_TORCH_OFF));
 
         reg(Material.SIGN, new ItemSign());
         reg(Material.REDSTONE, new ItemPlaceAs(Material.REDSTONE_WIRE));
